refactor: extract Ollama HTTP transport
